requirements.txt
tink
requests
brotli
pycrypto
zlib
urllib
pyyaml
ipaddress
python3-hvac